requirements.txt
torch>=1.10.0
transformers>=4.16.0
datasets>=2.0.0
scikit-learn>=1.0.0
seqeval>=1.2.2
PyMuPDF>=1.19.0
pytesseract>=0.3.8
opencv-python>=4.5.4
pillow>=8.3.2
numpy>=1.20.0
tqdm>=4.62.0